requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.15
langchain-core>=0.1.0
pydantic>=2.0.0
duckduckgo-search>=4.1.1
python-dotenv>=1.0.0 